+++ client/src/pages/Search.jsx
@@ -1,18 +1,18 @@
 import { useState, useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { useMutation, useLazyQuery } from '@apollo/client';
+import { useLazyQuery } from '@apollo/client';
 import Auth from '../utils/auth';
 import { googleBooks } from '../utils/API';
-import { SAVE_BOOK } from '../utils/mutations';
 import { GET_ME } from '../utils/queries';
-import { Form, Button, Card, Row, Col } from 'react-bootstrap';
+import { Form, Button, Row } from 'react-bootstrap';
+import SearchedBook from '../components/SearchedBook'
 
 const Search = () => {
   const navigate = useNavigate();
   const [searchInput, setSearchInput] = useState('');
   const [searchData, setSearchData] = useState('');
   const [savedBookIds, setSavedBookIds] = useState([]);
-  const [saveBook] = useMutation(SAVE_BOOK);
+
   const [getUser, { loading, error, data }] = useLazyQuery(GET_ME);
 
   //if user does not have valid token in their local storage, return them to the homepage
@@ -65,37 +65,7 @@ const Search = () => {
     }
   };
 
-  //this function saves a book to db and hides the div
-  const handleBookSave = async (event, book) => {
-  //validate user by getting jwt token from local storage and checking if it is expired
-    const token = Auth.getProfile() || null;
 
-    if (!token) {
-      return false;
-    }
-    try {
-      //check if book being saved is already saved
-      if (savedBookIds.includes(book.bookId)) {
-        console.log('already saved in database!');
-      } else {
-        const { data } = await saveBook({
-          variables: { book },
-        });
-
-        if (!data) {
-          throw new Error('something went wrong!');
-        } else {
-          //hide disable save button and hide book to prevent being saved twice
-          const card = event.target.parentElement.parentElement.parentElement;
-          card.classList.add('hide');
-          const button = event.target;
-          button.disabled = true;
-        }
-      }
-    } catch (e) {
-      console.log(e);
-    }
-  };
 
   return (
     <>
@@ -138,38 +108,9 @@ const Search = () => {
           </Form>
           {searchData && (
             <Row className="justify-content-center">
-              {searchData.map((book) => (
-                <Col key={book.bookId} className="mb-4" md={4}>
-                  <Card style={{ width: '18rem', paddingTop: '15px' }}>
-                    <Card.Img variant="top" src={book.image} />
-                    <Card.Body>
-                      <Card.Title>
-                        Title:
-                        <br />
-                        {book.title}
-                      </Card.Title>
-                      <Card.Text>
-                        Authors:
-                        <br />
-                        {book.authors}
-                      </Card.Text>
-
-                      <Card.Text>Pages: {book.pages}</Card.Text>
-                      <Button
-                        onClick={(e) => handleBookSave(e, book)}
-                        variant="primary"
-                        style={{
-                          cursor: 'pointer',
-                          fontFamily: 'IM Fell DW Pica',
-                          fontSize: '25px',
-                          backgroundColor: 'gray',
-                        }}
-                      >
-                        Save
-                      </Button>
-                    </Card.Body>
-                  </Card>
-                </Col>
+              {searchData.map((book, i) => (
+                //inject componenet passing props
+                <SearchedBook key={i} book={book} savedBookIds={savedBookIds} />
               ))}
             </Row>
           )}
